# Add short test mode to skip long-running tests