--- a/tests/test-cases/demo-vendoring.yaml
+++ b/tests/test-cases/demo-vendoring.yaml
@@ -2,6 +2,7 @@
 
 tests:
   - name: atmos vendor pull
+    short: false
     enabled: true
     snapshot: false
     tty: true
@@ -27,6 +28,7 @@ tests:
       exit_code: 0
 
   - name: atmos vendor pull no tty
+    short: false
     enabled: true
     snapshot: false # We can't use snapshots because temp paths will always be different
     tty: false