--- a/tests/test-cases/schema.json
+++ b/tests/test-cases/schema.json
@@ -15,6 +15,11 @@
               "type": "boolean",
               "description": "Whether the test is enabled or not."
             },
+            "short": {
+              "type": "boolean",
+              "description": "If false, skip this test when running with -short flag. Use for tests taking >2 seconds (network I/O, Git operations, heavy processing). Defaults to true.",
+              "default": true
+            },
             "tty": {
               "type": "boolean",
               "description": "Whether to run the test with a TTY."